refactor: split api/app.py into focused modules
